Feature: Attempting to create a propagator from the TLE object.
Summary: TLEPropogator is an abstract class, so I would have to instantiate either the SPG4 or DeepSPD4 class. Looking in the orekit-test classes shows that there is a different method for retrieving the propagated orbit for a satellite based on a different data source; a RINEX file. See Orekit-Tutorials estimation.GNSSOrbitDetermination and propagation.GNSSPropagation.